requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.13.0=py38h06a4308_0
aiohttp=3.7.4.post0=py38h7f8727e_2
async-timeout=3.0.1=py38h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
bottleneck=1.3.2=py38heb32a55_1
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2020.10.14=0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2020.6.20=py38_0
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=py38h06a4308_1003
charset-normalizer=2.0.7=pypi_0
click=8.0.3=pypi_0
coverage=5.5=py38h27cfd23_2
cryptography=3.4.8=py38hd23ed53_0
cudatoolkit=11.1.74=h6bb024c_0
cython=0.29.24=py38h295c915_0
dataclasses=0.8=pyh6d0b6a4_7
datasets=1.15.1=pypi_0
decorator=4.4.2=py_0
dill=0.3.4=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.3.1=pypi_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.11.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.36.1=py38h2157cd5_1
huggingface-hub=0.1.2=pypi_0
huggingface_hub=0.0.17=py_0
idna=3.3=pypi_0
importlib-metadata=4.8.1=py38h06a4308_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
jedi=0.18.0=py38h06a4308_1
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=11_linux64_mkl
libcblas=3.9.0=11_linux64_mkl
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=h14aa051_19
libgfortran4=7.5.0=h14aa051_19
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=11_linux64_mkl
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.1.0=py38h27cfd23_2
multiprocess=0.70.12.2=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
nltk=3.6.5=pypi_0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.3=py38h8c16a72_0
parso=0.8.0=py_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
prompt-toolkit=3.0.8=py_0
protobuf=3.17.2=py38h295c915_0
ptyprocess=0.6.0=py38_0
pyarrow=6.0.1=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.5=h7579374_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.10.0=py3.8_cuda11.1_cudnn8.0.5_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
readline=8.1=h27cfd23_0
regex=2021.10.21=pypi_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
sacremoses=0.0.46=pypi_0
scikit-learn=0.23.2=py38h5d63f67_3
scipy=1.5.3=py38h828c644_0
sentence-transformers=2.1.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py38hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=pypi_0
torchaudio=0.10.0=py38_cu111
torchvision=0.11.1=py38_cu111
tqdm=4.62.3=pypi_0
traitlets=5.0.5=py_0
transformers=4.12.0.dev0=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=py_0
werkzeug=2.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
xxhash=2.0.2=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
